requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
fastapi==0.112.1
h11==0.14.0
idna==3.7
influxdb-client==1.45.0
Jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==2.1.0
passlib==1.7.4
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
reactivex==4.0.4
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
